refactor: 랭킹 배치 서비스를 공통 배치 서비스 패키지로 이동
